Move components into folders
