+++ src/client/GlobalState.js
