+++ src/client/components/elements/imageElement.component.js
