+++ src/client/components/filters/addFilter.component.js
