+++ src/client/components/filters/filterView.component.js
