+++ src/client/components/jsonView/jsonView.container.js
