+++ src/client/components/rate/rate.component.js
