+++ src/client/components/search/searchResult.component.js
