+++ src/client/components/svg/svg.container.js
