+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from './client/App';
 import './css/index.css';
 
 ReactDOM.render(
